[location] cache and load region code
This change adds a cache for regulatory region services to store
the region code when it is configured. When the regulatory services
start up, the initial value is loaded from the cache.

The cache may be deleted if disk space is almost out. Regulatory region
services does not rely on this data to necessarily be saved; it still
expects to be configured at startup if needed.

Bug: 66537
Test: fx test regulatory-region-testst
      fx test regulatory-region-integration-test
      manually tested that region code is loaded at startup after
      rebooting.
Change-Id: Ic4ab13c6c6b6e1be4419ba6870e67783bd02e3e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471023
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>