[fdf] Clean up ETH_PHY platform defintion
This value is only used for Realtek drivers and it was put in the
fucshia.realtek.platform bind library, so we should make it clear that
it's for Realtek in platform-defs.h too. Also change the value to 0x03
to line up with the other Realtek DIDs.

Change-Id: If529954ab45ad54bab9a6e7dfaeff90c1750bc51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472521
Fuchsia-Auto-Submit: Alex Legg <alexlegg@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>