[driver_manager] Add support for loading fallback drivers sooner.
Currently when booting from a USB, the system hangs during early boot
because the `usb_composite` driver is a fallback driver, which means it
won't get bound until system drivers have been loaded - but system
drivers are on a USB device, which requires the `usb_composite` driver
to be bound.

This CL adds a flag, `devmgr.bind-eager`, which is a comma-separated
list of fallback drivers to treat as non-fallback.

Test: fx test ddk-fallback-test
Test: Boot from USB on NUC no longer hangs before minfs/blobfs comes up.
Bug: 44586
Change-Id: I9cef1954007a32ae862e294ac09de7fdd0afe89d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471837
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>