[fint] Check whether build graph is affected
This is effectively a translation of the should_build.py script from
the recipes repo into Go:
https://fuchsia.googlesource.com/infra/recipes/+/90d1d2a99f14/recipe_modules/build/resources/should_build.py

fint will now run the same `gn analyze` command after running `gn gen`
(although only on builders where build graph analysis is enabled). If
the analysis reveals that the changed files do not affect the build
graph, fint will set the `skip_build` field of the emitted artifacts to
`false`, so eventually recipes can check that boolean rather than
running should_build.py.

I made a few simplifications to the code compared to the should_build.py
script:
- Pass paths to input/output files as positional arguments rather than
  using stdin/stdout. This is the easier option for fint because
  the subprocess runner library doesn't currently support passing stdin
  to subprocesses.
- Don't bother deserializing some `gn analyze` output fields that we
  don't need, like "test_targets" and "invalid_targets".

This is effectively a no-op for now (unless the analysis fails for some
reason) because the result is ignored by recipes. It will add a couple
seconds to the build unnecessarily, but I think that's worth it to be
able to make a simple soft transition.

Before switching recipes to read this value instead of invoking
should_build.py, I'll add some temporary code to assert that fint and
should_build.py agree and leave that running for a few days to gather
data.

Bug: 65899
Change-Id: I1215b8a2dcabcb33a9f40f9e5111dc5a4d6975de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471117
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>