[spi] Add VMO registration to SPI protocol
SPI transfers are currently limited to 4096 bytes by the fragment proxy
driver. To allow for larger transfers, and to eventually enable DMA,
this change adds a methods to the SPI protocol that allows VMOs to be
used instead of Banjo vectors.

This change also adds a method to register a channel for making FIDL
calls using the SPI protocol. The fuchsia.hardware.banjo protocol can
be deprecated as soon as all drivers and clients have been moved to the
FIDL protocol.

Bug: 67570
Test: Built all SPI drivers and tests
Change-Id: Id6933a903d46b7d3d5c3186ebe1f9ba173ae4d32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470337
Commit-Queue: Braden Kell <bradenkell@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>