[docs] Freshen up docs/concepts/framework
Rename framework to process, to create a distinction of usermode
framework things that are ancestral to components.

Remove boot_sequence.md, it is hopefully outdated and the topic is
better covered anyway in
everything_between_power_on_and_your_component.md. Plus users have been
complaining that the stale contents are confusing to them.

Deleted hub as it's a components concept, the page contents were out of
date and not super helpful anyway, and the details vary between v1 and
v2 components so it's irrational to introduce this concept at this
point.

Deleted some links from markdown that's not under docs/. We don't care a
whole lot about markdown that's not on fuchsia.dev these days. Relics of
bygone days.

Change-Id: Iadab82ea63396f728b4684816ea7d5328da0a4d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472914
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Gary Bressler <geb@google.com>