[bt][a2dp] Initiate streaming if the peer doesn't
If the peer connects to us and doesn't start the stream, try to start
the stream using our normal method.

The delay is the same as the initatator delay.
Skip the delay when we establish the control channel.

Fixed: 50465
Test: added tests; fx test bt-a2dp-lib-tests
Test: connect to device manually and stream w/phone
Change-Id: I588fbb7f825535e11fc61cb088cd7d6041006b76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471462
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>