[boards][x64] Reduce differences between x64/qemu-x64/chromebook-x64.
In preparation for factoring out commonalities between these platforms,
attempt to reduce the differences in configuration between the x64,
qemu-x64, and chromebox-x64 board configs. Most differences appear to be
inadvertent.

While some differences are reasonable (such as chromebook-x64 not having
display drivers for "amd-kaveri"), we don't have a short term need to be
space-conscious on x64, and keeping the boards as similar as possible
reduces the maintenance burden.

Bug: 67913
Change-Id: Ia378a440c6cac0a3bb0311151a8384b71c76097d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472701
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: David Greenaway <dgreenaway@google.com>