[docs] update kernel debugging documentation
I updated the instructions to connect to use gdb to connect to qemu.
The old instructions didn't mention that the usage of the gdb script
was mandatory (for kaslr relocations). The new documentation
explictly mentions that you should use the script for using remote gdb
then it mentions that it is possible for it to load automatically (if 
set up correctly), and notes tha it won't work right away with qemu-kvm

BUG=67893

Change-Id: Id4159e74f9f109b91039465d59bb0ef730c52cca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472417
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Tucker <raggi@google.com>
Fuchsia-Auto-Submit: Marco Vanotti <marcovanotti15@gmail.com>