[vfs] Add initial pager capability to C++ VFS.
This adds the initial skeleton of pager capabilities to the C++ VFS. The
long-term goal is to replace blobfs' pager framework with this one, and
the even longer-term goal is to remove the non-pager paths from the VFS.

This patch implements the simplest possible read support and a test that
runs against it. There is a lot of work left, especially around closing
files and shutdown. There are also a lot of needed test cases left, some
of which are listed at the bottom of the pager_test.cc file.

Bug: 51111

Change-Id: I17e427db68b4fcf30dfebbe18117e586ef8f0cc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466474
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>