[camera][lib] image_io_util allows caller to specify dir_path
This CL changes the util's behavior to allow callers to specify the path
to write frames to, replacing the previous hardcoded solution.

Example usage of the util can be found in the unittest, which writes to
the on-device cache.

Change-Id: I17cdcc1fc912476ef8cdfe212c00ace9c539ae38
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471637
Commit-Queue: Nic Zoghb <nzo@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Fuchsia-Auto-Submit: Nic Zoghb <nzo@google.com>