[fidl][hlcpp] Split golden tests
This splits the HLCPP golden tests into two tests, one for the
traditional bindings and one for the "natural types" variant. This
allows the goldens for a given test to all be in the same directory
which makes the build rules (and the changes I'm trying to make to them)
much saner.

Change-Id: I1db49d5d726e93ef73cce7a90a6982b9727a9590
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471318
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>