[httpsdate] Only record offset for the second poll
HTTPSDate attempts to schedule each polls such that the bound on UTC
time it has is halved by the poll. This is achieved by scheduling the
poll such that either the minimum or maximum UTC possible with the new
poll falls in the center of the accumulated bound. After a few polls,
the accumulated bound is very small. When HTTPSDate schedules another
poll in this situation, the center of that poll is very close to +/-
.5 second and skews our metric.

This change updates HTTPSDate to record only the second poll to remove
this effect. The first poll is not used as it usually takes longer due
to setting up TLS connections. The second poll is likely still fairly
random as the algorithm has not yet tightened the bound.

Bug: 63370

Change-Id: I401183efdc503bd0352a6ec2c69bb3b15f131ff3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472217
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Jody Sankey <jsankey@google.com>