[rust] Add type for padding in zircon structs.
From the earlier discussion in fxr/468595, some zircon C structs contain
manual padding bytes to control alignment. The content of these bytes is
not specified by contract hence they should not be considered when
testing for equality.

Prior to this CL, in rust we copied these padding bytes to guarantee the
same alignment as the C structs, but the derived PartialEq and Eq traits
were considering the padding content when testing equality.

This CL defines a new `PadByte` wrapper for u8 whose PartialEq, Eq,
Hash, and Debug implementations all ignore the content of the byte. All
struct field named padding* have been replaced with this types, and a
new test verifies that two structs are considered equal even if the
padding content differs.

Change-Id: I75a6b29d711bbba9c32b390d61786f132b6662eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471200
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>