[kernel][vm] i-cache consistency manager
Adds a per-arch i-cache consistency manager to efficiently sync
multiple ranges. Although the only place we presently sync multiple
ranges is in VmObject::CacheOp, more will be added in later commits.

The ARM implementation of the consistency manager conforms to the
pessimistic VIPT i-cache, but can be trivially extended to check
bits in the CTR to optimize. Actually doing this is deferred as
all current primary platforms are VIPT.

Change-Id: I85c11e6e57eb75feff001929cae1d46a9558b29b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466060
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>