[fidl_codec] harden handle tests
To test fidl codec, we send a message over a channel and decode the
message received at the other end.

For channels, the driver can change the numerical value of the channel
(and it will change it every time if a restriction of rights is asked).

This patch ignores the numerical value of the received handles and
replaces it with well known values. This way, we can reliably check the
decoded message text.

Change-Id: I846519578ca88fc5d46deb39d9e9de3910d8b7f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473332
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Vincent Belliard <vbelliard@google.com>