[kernel][x86][topology] Address cosmetic bug
Today we enumerate dies using the tail of the APIC ID that actually
encodes both the canonical package ID and die ID. Although this ID
results in the same effective enumeration of the dies in the system -
in comparison to enumerating all packages using package ID and in turn
all dies within by die I - it gives a misleading presentation both in
source (in the definition of the APIC ID decoder) as well as at runtime
in the early boot dump of the system's topology.

To address this, we introduce a `Package` abstraction to the topology
enumeration process and update the `ApicDecoder` to identify both
packages and dies. This is a purely cosmetic change to the enumeration
internals and results in the same system_topology::Graph being
constructed.

Bug: 61093
Change-Id: I3875257c66040820e4d2f53204a7c74596a8e1dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465897
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>