[kernel][code-patching] Initial commit
This change introduces a new code-patching library, which offers support
for recording code-patching metadata in assembly, a utility to nop-fill
an instruction range, and a GN template for extracting the raw section
content where the code-patching metadata is stored.

We also introduce a self-modifying phys ZBI test that exercises the
nop-fill case.

Bug: 67615
Change-Id: I149f7ce20ab94e554673539039c0515616236b03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470662
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>