[ping] Fix IPv6 scope id override
Only overwrite the IPv6 scope id if a value has been provided with the
`-I` argument and it is positive. This was preventing specifying the
scope id after a % in the URL from working, as `getaddrinfo()` would
populate the scope id field correctly but it would be overwritten with
a value of 0 unless `-I` was provided.

Fixed: 64465
Change-Id: I77bc380f49dc3e06e0485c00dc94c06c07386514
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/452660
Reviewed-by: Tamir Duberstein <tamird@google.com>
Testability-Review: Tamir Duberstein <tamird@google.com>
Commit-Queue: Tony Gong <gongt@google.com>