[kernel][sancov] fix kernel sancov variant.
It seems like sometime during the build unification, the sancov variant
bit rotted. The configs were renamed and the sancov rule was not
updated, as it's not being covered by any buildbot.

This CL fixes that, setting the config to the correct value for the
sancov and asan configs.

Test:

```
 $ fx set core.x64 --variant "kasan-sancov" && fx build
```

BUG=68034

Change-Id: Iddbac6a28706d7e140cd0fb9d17684b373c83ae0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473286
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Marco Vanotti <mvanotti@google.com>