[reachability] Fix flake due to inspect race
Due to the non-atomicity of inspect entry creation, when reachability
monitor adds an inspect entry for reachability state into the
`BoundedListNode`, the sequence of events is as follows:

1. an entry is created with a sequence number only,
2. the timestamp is written into the entry, and
3. the reachability state as a String value with the key "state" is
   written into the entry.

This means that it's possible to observe an entry that does not have
the reachability state string value in it.

The integration test was flaking because it treated tho lack of the
"state" string value as an error condition. This has been fixed by
assuming the IPv4 and IPv6 reachability state to be "None" if the
state value cannot be found in the inspect data.

Fixed: 67541
Test: fx test netstack_reachability_integration_test \
              reachability-core-lib-test
Multiply: netstack_reachability_integration_test:30
Change-Id: I100d77a7579eb26f1f829db32cf342825d239249
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473198
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>