[component_manager] Rename Realm -> ComponentInstance
`ComponentInstance` is a better name for this struct because at its core
it represents one ComponentInstance. It _is_ a realm too in the sense that
it contains links to children, but that notion is somewhat secondary.

Also rename related modules, identifiers and variables.

Change-Id: I958ea85898d6dd97a45a7e997c704a9a81115757
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472746
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Gary Bressler <geb@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>