[settings] Implement transforms for intercepted changed events
When the audio setting's value changes in response to a Set
call, it sends a Changed event to the switchboard to tell
it to notify clients listening over a hanging get of the
new value.

This CL captures and modifies the Changed event in the
audio policy handler so that Set requests result in
listeners receiving the properly transformed volume levels
based on the active policies.

Fixed: 67148
Test: fx test setui_service_tests
Change-Id: I9895754d6ef340c68d551b04e4e762a546e18383
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471142
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>