[weave] Implement conditional TSM init
ThreadStackManager will now check the configuration value
("enable-thread") before initializing. If Thread support is not enabled
in configuration, TSM now exits without failing when a Thread device is
not found, but any further calls to TSM will not succeed (either
returning an error indicating the functionality is not supported or, for
those calls that don't return a WEAVE_ERROR, returning false or other
appropriate value and provide a guarantee that there will be no side
effects to the call).

Bug: 67672
Test: fx test weavestack-adaptation-unittests
Change-Id: I13962c682078a9d22d53863c47c11d169c789539
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471617
Commit-Queue: Jason Graffius <jgraff@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>