[brcmfmac] Fix spurious failures in token bucket
The token bucket could sometimes allow additional tokens to be consumed
if the atomic compare exchange operation spuriously failed. This would
then load the old tick and not perform the necessary checks to ensure
the bucket did not go over capacity.

This rarely triggered a test failure, added another test to run for an
extended period of time.

This fix ensures that the capacity is properly checked on each attempt
to consume a token.

BUG: fxbug.dev/68044
Test: Included unit test
Change-Id: I9099b0b8cd6e78404d10ac481fcf198ad6c430f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474197
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>