[analytics] Executor with thread and timeout
This change add a new implementation of fit::executor, which manages a
debug_ipc::MessageLoopPoll in a separate thread. It also sets a quitting
timeout, such that when the main thread quits, the message loop will try
to quit itself after the timeout if tasks are not finsished earlier.
Note that the timeout might not work if there are blocking tasks
running on the loop.

The analytics_executor_manualtest is an integrated test to make sure
libcurl works well with the threading and timeout logic implemented in
AnalyticsExecutor.

MULTIPLY: host_x64/zxdb_tests: 50
Bug: 65915
Change-Id: I3291de6dc7e659057ae045980bf6bd14f16bc9ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465244
Commit-Queue: Rundong Du <rundongdu@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>