[camera] Force SetCropRect call in all cases
Also augmented unit tests to check that the call is made.
Added unit test to make sure call is made even without Start().

The basic problem with b/177099834 is that it is entirely possible to
send a SetCropRegion request before that particular Stream is ready
to flow. What happens is that the request is sent to DeviceImpl, and
remembered as the last crop region requested. Just before starting the
stream, StreamImpl will push down that saved SetCropRegion request as
a SetRegionOfInterest to the camera controller.

The problem is that this last moment request assumes that the
SetRegionOfInterest request will actually take place even if the
stream isn't enabled. Unfortunately, the current implementation of
Ge2dNode does not allow the request to go thru to the GE2D driver if
the Stream is not enabled (starting the Stream will enable it). The
effect is that DeviceImpl will send the SetRegionOfInterest request
just before calling Start, and the SetRegionOfInterest will be ignored
by the controller.

The fix presented here is to go ahead and push down the request to the
GE2D driver anyway, even if the Stream is not enabled. What should
happen is that the call to GE2D is always done unless there is some
parameter error discovered along the way.

Bug: b/177099834
Test: fx test camera-internal-tests //src/camera
      Also force failures to make sure checks are really checking the
      right things.
      Also run same unit tests against old code to ensure the key
      check fails.
Change-Id: Ic01319080f9207bb335325f8185e52864cf176ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471599
Commit-Queue: Ernest Hua <ernesthua@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>