[stdcompat]: Proper exception handling.
Mistakenly used 'assert' for well defined behavior of the polyfill.
Fuchsia builds without exceptions enabled, but sdk user might have them
enabled, and the polyfill should conform to the standard.

Moved to use [[gnu::unused]] for C++14(clang and gcc have support for
this attribute).

Also added asserts for undefined behavior in string_view, this will only
be triggered on !NDEBUG builds, meaning that are stripped out on release
builds, hopefully we can catch errors earlier.

test_helper.h provides ASSERT_THROW_OR_DEATH, which depending on
whether exceptions are enabled or not will either catch an exception and
verify that is the right type(matches spec) or a death test that
verifies program execution is aborted.

For checking that on debug builds of our polyfills, asserts remain so we
can catch bugs early, DEBUG_ASSRET_DEATH is a GTEST_SKIP unless we are
enforcing use of polyfills on a !NDEBUG build.

Bug: 57349
Test: stdcompat-tests
Change-Id: I3fae2b470eac27f225483206aee6d33f8137802b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472438
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>