[kernel][arch] Add APIC ID utilities
We add utilities for fetching the APIC ID of the processor associated
with a CpuidIoProvider, as well as for partitioning a given APIC ID into
its consituent thread, core, die, and package IDs. This will be used in
a follow-up in a refactor of the topology enumeration in
//zircon/kernel/arch/x86/system_topology.cc. Shortcomings of the code
there today includes:
* offers fewer fallback means of enumeration
* does not look at leaf 0x1f for the V2 means of enumeration, which
should take precedence of leaf 0xb;
* does not check for the "topology extension" feature before consulting
AMD's topology leaves (which are reserved if not advertised);
* does not allow for zero-width IDs;
* does not distinguish between die and package IDs;
* does not make use of x2APIC IDs (only the initial 8-bit one of the
enumerating processor);
This and more is addressed in the new abstractions, and is tested
against the corpus.

Bug: 61093
Change-Id: I4afbf229f280f746392bac2c61c7f2386b6b2462
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465058
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>