[ulib][zxc] Make zxc_unittests runnable on host
There is one test for zx::status that can't run on host because of a
missing member function, but it's very useful to have and much faster
for your edit-compile-debug loop to just go into
out/default/host_x64/exe.unstripped and run the zxc_unittests binary.
It even works with regular gdb so you don't have to attach zxdb to an
emulator.

Of course there is some testing that needs to be done on Fuchsia but the
vast majority of zxc is platform-independent.

TEST=ran the newly generated binary, tests passed.

Change-Id: I3e83cf9088fd782b069a9b9d692fd0a71d07703a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474409
Commit-Queue: Mark Schott <schottm@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>