[scrutiny] BlobFS enable support for Version 9 compact merkle trees.
BlobFS version 9 introduces compact merkle trees at the end of the
extend instead of as a prefix. This saves a lot of space but also
changes the format of how data is stored and how the merkle tree size is
calculated. This change enables support for postfix merkle trees and
supports both V8 and V9 formats.

Change-Id: Ie95517dbc959926385e4e96000dfb6acc80b8055
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473497
Reviewed-by: Yu Shan <shanyu@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>