[kernel][vm] Don't panic on DECOMMIT w/ multiple vmar mappings
A VMAR may have multiple  mappings inside its address range.
If a VMAR had multiple non-contiguous mappings and the decommit
vmar op was invoked on mappings that weren't the first one,
the DECOMMIT lambda in kernel would panic, noticing it was
called on a mapping that didn't include the address to decommit.

The bug is because the iterator EnumerateChildrenInternalLocked
walks a vmar's subregion tree and invokes on_mapping on every
mapping in a subregion, not just ones that contain the address
in question. Fix the bug by checking that the address is contained
by a mapping before invoking on_mapping.

Pair-programmed-with: mvanotti@
Bug: 68272 ASSERT FAILED at (../../vm_address_region.cc:638)

Change-Id: I460482104fa6144d7e16f3d9542b3cd4f1e6901f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474719
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>