[lib][vmo_store] Make VmoStore return unregistered VMOs
In some cases it may be helpful to operate on VMOs after unregistering
them, or be able to pass them back to a client of the driver. Currently
Unregister() just closes the VMO handle; this change makes it so the VMO
is passed back to the user instead.

Fixed: 68209
Test: fx test vmo-store-test
Change-Id: I16ca1efee7fb900fc50a0e5ee0292b5aa54951b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474700
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>