[scrutiny] Add v2 component tree as a DataCollection
Constructs a ComponentTree from the set of all v2
components and retains it as a DataCollection.
Also adds a placeholder plugin `verify.map_tree`
which just lists all nodes of the tree in
breadth-first order.

Changes ComponentTreeBuilder's error handling: now,
instead of returning an error immediately if a
ComponentDecl is not available for some child URL,
the builder constructs as much of the tree as
possible. The builder's new return value now
contains both the tree (if any) and a vector of
all errors encountered.

In this CL, Scrutiny just logs each of these errors
as a warning, but a future plugin could operate on
them.

Change-Id: Ib843d5433f7f747add14e98342265b316ff6a961
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475138
Commit-Queue: Benjamin Wright <benwright@google.com>
Fuchsia-Auto-Submit: Laura Peskin <pesk@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>