[fx][qemu] Add an --archive option to the qemu command
Add a --archive options to the qemu command which can be used to
create a compressed tarball with all of the files needed to launch the
VM, instead of actually running the image.  This option can be helpful
when someone wants to take a build over to a remote machine (such as a
Cavium workstation) and run it there instead.

Change-Id: I69215897e62b1562073a067e2e7012b225c94e6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474254
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>