[i18n] Hermeticize static_icu_data build template
Added explicit references to input files in this build rule.  While
technically we still read the entire directory, in practice the contents
of the directory never changes - it's always the same files.  So it
should be OK to name the only 3 files that are currently there for the
time being.

Bug: 68703
Change-Id: I1f0048fd0b07b2980497d72cd43e6cd1dade68d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475179
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>