[stdcompat]: Variant exception cleanup.
Change the tests to use ASSERT_THROW_OR_ABORT, to work as well with
enabled exceptions as disabled.

cpp17::varinat is current missing a key feature, which is the ability to
use an equivalent of cpp17::visit. This should be provided as well,
since provide the ability to do automatice static dispatch.

Test: stdcompat-tests
Change-Id: I251ef2f18128ae6fbf17d4c2e9d51c5dfee64202
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472906
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>