[harvester] Listen for ZX_CLOCK_STARTED
This introduces a FuchsiaClock class which returns an optional for the
now value of time. We can use this to determine if the system clock has
started or not. We then use this class to supply times to the various
RPC messages that need times. This class has a mechanism for calling a
callback when the clock starts and this is used to send the ClockStarted
RPC to Dockyard.

I haven't added tests yet but wanted to put this up for feedback before
I finish that.

Change-Id: I617886aae740b1834a7cc11e78941f14d5597373
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/473862
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Ryan Knauer <rknauer@google.com>