[settings] Migrate mic to input device
Migrates the old format of microphone to be in the new format for input
devices. Internally it will now use the input device format when a
client toggles the mic state. The ability to toggle with the old state
should be removed once all clients are migrated to use the new
interface.

Fixed: 67211
Test: fx test -o setui_service_tests

Change-Id: I13fc2bb737dac36d4eb78e630ca84c60305e9c42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474113
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>
Reviewed-by: William Xiao <wxyz@google.com>