[docs][fx]
Revised the intro to "What just happened?"

It was a little unclear as written.

1. removed the first bullet, which didn't correspond to a separate part
of the example `fx set` command and was redundant with the `board` bullet.
2. Added an intro to the discussion about `base`, `cache`, and
`universe` because these terms are discussed without relating them to
anything in the `fx set` example or anything else above.

Change-Id: Ia4f1eeabfbb68146411a70f426f7130b7ca7ad19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474750
Commit-Queue: Rich Kadel <richkadel@google.com>
Reviewed-by: Kelly-Ann Dolor <kadolor@google.com>