[modular] Serve outgoing dir in FakeComponent after OnCreate
Ensure that the outgoing directory is only served after all
services are added to it, as typically this is done inside OnCreate.
This is correct behavior but effectively a no-op because it happens
in the same async loop frame either way.

More importantly, this enables tests to perform tasks between
creating the component and serving the outgoing directory, useful
for testing racy code that connects to services.

Change-Id: I33e0ddd88d773a388ab94013b3da68eff35eb6fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475743
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>