[time][flake] Disable timekeeper unit test on emulators.
We're still getting a persistent low rate of flakes on emulators for
tests that measure any flavor of time accuracy. This is primarily
since timekeeper can be prempted for many hundreds of milliseconds
between calculating times and the kernel using these times.

Once we have a zircon API that doesn't depend on time of delivery it may
be possible to re-enable these tests.

Bug: 67711
Change-Id: I06fd0ddbdacb5fca2ae93577a2359010b3a7e166
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/474749
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>