[blobfs] Add destructor ordering shutdown test for pager.
This test intentionally blocks the pager thread while it is handling the
page fault so that the pager can be shutdown while it is handling a
page. This will induce a crash on the pager thread if it attempts to use
any member variables that are ordered to destruct before the Async::Loop
attemps to run Shutdown(), as the Shutdown() method will await worker
completion.

An important point here is that the page fault may still fail despite
proper shutdown ordering. The pager may detach as part of destruction
before some post-fulfillment checks in the kernel verify that it is
still attached and return failure to the faulting thread anyway. For
this reason, the success of the zx_vmo_op_range() call is not checked.

MULTIPLY: blobfs-unittest (fuchsia):20

Change-Id: I41ff40b54a2b58d81efce95af31e7ff222ac8851
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472844
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>