[kernel] Unittests of kernel stack depth handling
Add a unittest that checks that we can receive interrupts and handle
mp_sync_exec calls when thread context has used up half of the kernel
stack.

The first test is simple and probablistic - it allocates a large
on-stack buffer and waits a bit for any (potential) interrupts. If an
interrupt were to take > 1/2 the kernel stack to handle, we would see an
architecture-specific stack overrun failure.

The second test is deterministic and ensures we can run an mp_sync_exec
handler (that signals an event) while half the kernel stack is used from
thread context by a buffer.

Bug: 64808 Panic on boot on the kasan bringup builder
Change-Id: I40d356197d38e039d999fbe09615921b08ab0a88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461084
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>