[virtualization] Disable bridging until netstack is fixed
Using bridging currently causes the netstack to crash, or an error to be
returned if promiscuous mode can not be set on a network interface.

Disable virtualization's use of bridging until it is fixed in the
netstack.

Bug: 67566
Change-Id: I10eff17e4ff2f27f2b177dfd3bfd04cc63c0e700
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476119
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>