[console-launcher] add /hub
The hub directory was removed from the serial shell some time ago,
because due to a CFv2 migration requests into this directory would hang
on bringup and this broke the shell.

Later the component framework was configured to disable the fuchsia-pkg
resolver on bringup, which solved this issue, so the directory can be
safely reintroduced.

Fixed: 50805
Change-Id: I3a408bd60c5541ccbddd72ea8338a2a690c80456
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476258
Fuchsia-Auto-Submit: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>