[system-update-committer] implement wait_and_reboot
This change adds reboot functionality to the system-update-committer.
Generally, if we fail to put the boot metadata in a happy state, we'll
reboot. Exception: if it's a blobfs verify failure, and the config says
to ignore, we won't reboot.

Note: we can't yet integration test the exception because currently the
blobfs verification will always pass.

Fixed: 64600
Test: fx test -o system-update-committer-tests
Test: fx test -o system-update-committer-integration-tests
Change-Id: Ic48be5220870c31ab0450cd2ea19f46ddaef6021
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/475560
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>