[build] Make host_test_data() play nicely with variants
For a build with variants of the host toolchain, let
VLT (Variant Location Test) = the variant-selected host test (i.e., the
one that gets put it host_x64-asan)
BLT (Base Location Test) = a hardlink of the VLT to the "normal" build
subdirectory (i.e., host_x64)

host_test_data() is meant to record metadata of runtime dependencies of
a host test for Infra's benefit (i.e., so that it can grab those files
along with the associated test's executable, and then run everything
outside of the build directory). This change is meant to address the
current limitation of 'VLT/BLT of tests and their runtime deps', which
can mean a few things:

(1) Test metadata records the test's path as VLT or BLT, and the test's
runtime deps as one of the two possible locations it could have been
copied to; these can diverge. This is a general problem/bug, which can
be quite frustrating and confusing to hit on. We tidy up a previous
mitigation against this in this change.

(2) One of VLT/BLT executes successfully, while the other complains that
it can't find its dependencies.

(3) The BLT does not have dependencies on its runtime deps. In this
situation minimally building the BLT won't result in the runtime deps
being copied to the place it expects them.

Change-Id: I1e7a1228438a75b0938c1125022959c48191c070
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476198
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: David Turner <digit@google.com>