[wlan][policy] Allow SME layer to complete its re-connect attempts
When it detects a disassociation, the SME layer automatically attempts
to reconnect. This reconnect is visible within the `connecting_to`
field of the SME status response.

Prior to this change, the Policy layer ignored the `connecting_to` field
and performed its own reconnect logic. When the SME layer was already
reconnecting, this interrupted the reconnect and resulted in a longer
delay to get reconnected.

Bug: 65611
Test: fx test wlancfg-tests
Multiply: wlancfg-tests
Change-Id: I91ca7e0cca0235d4842ac2333c9f6c73cd5c1fb0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476519
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>