[brcmfmac] Remove ASSERT in brcmf_notify_start_auth()
This function used to ASSERT when seeing an AUTH_START event while the
iface is not in CONNECTING state. This scenario is verified to be
expected by ACTS test, when an WPA3 authentication is interrupted by an
deauth_req from SME before firwmare sends up the START_AUTH.

So the ASSERT should be remove and the log level should also be changed.

Test: None, no functional change.
Bug: 68540
Change-Id: I1a46502709d087e35ab83f6d873b93637a389022
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476937
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Zhiyi Chen <zhiyichen@google.com>