[wlan] Improve replay counter testing
Previously untested replay counter behaviors are added to the test
suite. These include the following tests.

- Ensure the first fourway exchange EAPOL message never increments the
replay counter. (It does not contain a MIC.)
- Add checks of the EAPOL response frame instead of assuming Ok()
indicates correctness. Based on the current state of the test suite,
Ok() is not always indicative the test succeeded.
- Assert the state of the key_replay_counter as appropriate.
- Ensure the third fourway exchange EAPOL messaeg increments the
replay counter.

There still remains a flaw in the replay counter implementation which
is that the MIC is not checked before incrementing the replay
counter. See http://fxbug.dev/68916

This change also renames send_msg1 and send_msg3 as send_fourway_msg1
and send_fourway_msg3 to clarify which exchange they are for.

Bug: 68916, 68390
Change-Id: I1a7fbde27c6ec3d1b24f658af491c1682bf3d6a2