[wlan] FIDL MLME: rename *Codes -> *Code, move ReasonCode to ieee80211
This is a functionally no-op change that renames the *Codes (e.g.
JoinResultCodes, AuthenticateResultCodes) to the equivalent Code (e.g.
JoinResultCode, AuthenticateResultCode), better matching the MLME
specification in IEEE 802.11-2016.  It also removes the ReasonCode
enumeration within the MLME FIDL and moves to using the ReasonCode
enumeration as defined in the ieee80211 FIDL.  In the process, some
usages of the ReasonCode that were previously typed as uint16 are now
explicitly typed as ReasonCode, for better semantics and type safety.

Bug: 29277
Bug: 55667
Test: fx test wlan-sim-env-test wlan-sim-fake-ap-test brcmfmac-sim-test
      wlanif_unittest wlan_tests wlan-mlme-tests wlan-rsn-tests
      wlan-sme-tests wlanstack-tests
      Booted and connected to WLAN on Astro

Change-Id: I9a2dc965be493e6206e61df448eecab87b8b2df7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/476697
Commit-Queue: John Sheu <sheu@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>